+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 1.0.1
 
-- Fix issue with setting `Circle#radius=` causing an infinite loop with very small decimals
+- Fix issue with infinite loop upon setting `Circle#radius=` to a number with very small decimals
+- Fix issue with infinite loop upon setting `Circle#diameter=` to a number with very small decimals
 
 ## 1.0.0